+++ frontend/components/sign-in-form.tsx
@@ -1,0 +1,18 @@
+import React from "react";
+import StyledFirebaseAuth from "react-firebaseui/StyledFirebaseAuth";
+import firebase from "firebase/compat/app";
+import "firebase/compat/auth";
+
+const uiConfig = {
+  signInFlow: "popup",
+  signInOptions: [firebase.auth.EmailAuthProvider.PROVIDER_ID],
+  signInSuccessUrl: "/",
+};
+
+const SignInForm = () => {
+  return (
+    <StyledFirebaseAuth uiConfig={uiConfig} firebaseAuth={firebase.auth()} />
+  );
+};
+
+export default SignInForm;
