+++ frontend/pages/sign-in.tsx
@@ -1,25 +1,9 @@
-// Import FirebaseAuth and firebase.
-import React from "react";
-import StyledFirebaseAuth from "react-firebaseui/StyledFirebaseAuth";
-import firebase from "firebase/compat/app";
-import "firebase/compat/auth";
+import SignInForm from "../components/sign-in-form";
 
 import "../utils/firebase/init";
 
-// Configure FirebaseUI.
-const uiConfig = {
-  // Popup signin flow rather than redirect flow.
-  signInFlow: "popup",
-  // Redirect to /signedIn after sign in is successful. Alternatively you can provide a callbacks.signInSuccess function.
-  signInSuccessUrl: "/",
-  // We will display Google and Facebook as auth providers.
-  signInOptions: [firebase.auth.EmailAuthProvider.PROVIDER_ID],
+const SignInPage = () => {
+  return <SignInForm />;
 };
 
-function SignInScreen() {
-  return (
-    <StyledFirebaseAuth uiConfig={uiConfig} firebaseAuth={firebase.auth()} />
-  );
-}
-
-export default SignInScreen;
+export default SignInPage;
